(#2718) Render main navigation via Edge Side Includes when Akamai module is active.. - Move rendering of main navigation block to cgov_core.navigation_block service. - Update documentation referring to non-existant NavItemInterface interface. - Render the ESI tag via a specialized variant of the template for the cgov_main_nav block. - Create controller to render main navigation.
