+++ docroot/profiles/custom/cgov_site/modules/custom/cgov_core/cgov_core.services.yml
@@ -53,3 +53,6 @@ services:
     class: Drupal\cgov_core\CgovConfigOverrider
     tags:
       - {name: config.factory.override, priority: 5}
+  cgov_core.navigation_block:
+    class: Drupal\cgov_core\Services\NavigationBlockManager
+    arguments: ['@cgov_core.cgov_navigation_manager','@language_manager', '@logger.channel.cgov_core']
