+++ docroot/profiles/custom/cgov_site/modules/custom/cgov_core/src/NavItem.php
@@ -306,7 +306,7 @@ class NavItem {
    * Optional, pass an array of class properties
    * with boolean values to filter children against.
    *
-   * @return \Drupal\cgov_core\NavItemInterface[]
+   * @return \Drupal\cgov_core\NavItem[]
    *   Filtered array of direct descendents.
    */
   public function getChildren() {
