+++ docroot/profiles/custom/cgov_site/modules/custom/cgov_core/src/Services/CgovNavigationManager.php
@@ -332,7 +332,7 @@ class CgovNavigationManager {
    * @param string $testFieldName
    *   Name of field to check for root status.
    *
-   * @return \Drupal\cgov_core\NavItemInterface|null
+   * @return \Drupal\cgov_core\NavItem|null
    *   Nav Item representing nav root term.
    */
   public function getNavRoot(string $testFieldName) {
@@ -354,7 +354,7 @@ class CgovNavigationManager {
    * @param \Drupal\taxonomy\TermInterface $term
    *   Base Term to wrap.
    *
-   * @return \Drupal\cgov_core\NavItemInterface
+   * @return \Drupal\cgov_core\NavItem
    *   Nav Item wrapping given Term.
    */
   public function newNavItem(TermInterface $term) {
