+++ docroot/profiles/custom/cgov_site/themes/custom/cgov/cgov_common/templates/block/main_nav/block--cgov-main-nav-esi.html.twig
@@ -1,0 +1,14 @@
+{#
+  This template contains a placerholder for when the MainNav block will be
+  added to the page via Akamai's ESI capability.  Refer to MainNav::build()
+  and cgov_core_theme_suggestions_block_alter() in cgov_core.
+
+  Variables (provided by MainNav::build())
+  * block_id - the block being rendered
+  * ?????
+#}
+<nav id="mega-nav" role="navigation">
+  <ul class="menu nav-menu">
+    <esi:include src="https://www-int-ac.cancer.gov/{{ content.lang_path }}block/{{ content.block_id }}" />
+  </ul>
+</nav>
